Moved the configuration settings to 'config.py'
I moved the configuration settings for the database into a separate file in the hopes of better cleaning everything up, but what do I know?

It seems to me that it will work out to be just fine, but I'm still too new to understand all of the OOP aspects just yet.